Resolved conflicts:
- sanctum.php: (int) 캐스팅 버전 사용
- login.blade.php: safeRedirect() 보안 수정 버전 사용

+++ config/sanctum.php
@@ -47,7 +47,7 @@ return [
     |
     */
 
-    'expiration' => env('SANCTUM_TOKEN_EXPIRATION', 60 * 24 * 7), // 7 days (in minutes)
+    'expiration' => (int) env('SANCTUM_TOKEN_EXPIRATION', 60 * 24 * 7),
 
     /*
     |--------------------------------------------------------------------------
